fix api integration and improve ui
fixed api calls (add, update, delete tasks) with authorization token
handled token using sharedpreferences
fixed task priority issue with backend enum
improved task state updates in view model
improved home and details ui
added task editing (title, description, date, priority)
added delete confirmation dialog
improved date picker styling
fixed duplicate dates in filter
improved task item display and interactions

+++ lib/features/home/presentation/widgets/task_item_widget.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:todo/core/common/widgets/svg_icon_button.dart';
 import 'package:todo/core/constants/app_assets.dart';
+import 'package:todo/core/dialogs/app_dialogs.dart';
 import 'package:todo/core/theme/app_colors.dart';
 import 'package:todo/core/theme/app_text_styles.dart';
 import 'package:todo/features/home/presentation/model/task_model.dart';
@@ -91,9 +92,16 @@ class TaskItemWidget extends StatelessWidget {
               children: [
                 SvgIconButton(
                   imagePath: Assets.assetsIconsTrash,
-                  onTap: () {
-                    context.read<TaskViewModel>().deleteTask(task.id);
-                  },
+                  onTap: () async {
+                final confirm = await AppDialogs.showDeleteDialog(
+                  context,
+                  title: task.title,
+                );
+
+                if (confirm == true) {
+                  context.read<TaskViewModel>().deleteTask(task.id); 
+                }
+              },
                   color: AppColors.secondColor,
                 ),
                 Container(
